chore: convert src/middleware files to TS
